+++ src/background/schema/fallbackApi/fallbackInfo.ts
@@ -3,6 +3,7 @@ import zod from 'zod';
 export const fallbackInfoScheme = zod.object({
     vpnApiUrl: zod.string(),
     authApiUrl: zod.string(),
+    forwarderApiUrl: zod.string(),
     expiresInMs: zod.number(),
 }).strict();