+++ src/popup/actions/popupActions/index.ts
@@ -2,11 +2,6 @@ import { tabs } from '../../../background/tabs';
 import { messenger } from '../../../common/messenger';
 
 class PopupActions {
-    openRecovery = async (): Promise<void> => {
-        await tabs.openRecovery();
-        window.close();
-    };
-
     openTab = async (url: string): Promise<void> => {
         await tabs.openTab(url);
         window.close();