+++ tests/background/endpoints/locationsService.test.ts
@@ -15,6 +15,8 @@ jest.mock('../../../src/background/stateStorage', () => {
     return require('../../../src/background/stateStorage/mv2');
 });
 
+jest.mock('../../../src/background/config', () => ({ FORWARDER_URL_QUERIES: {} }));
+
 jest.mock('../../../src/background/connectivity/pingHelpers');
 jest.mock('../../../src/common/logger'); // hides redundant log messages during test run
 jest.mock('../../../src/background/settings');